% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_style.R
\name{oz_style}
\alias{oz_style}
\title{Add OZ theme to ggplot chart}
\usage{
oz_style()
}
\value{

}
\description{
Add OZ theme to ggplot chart
}
\examples{
# a basic histogram
data.frame(x = rnorm(1000)) \%>\%
ggplot(aes(x = x)) +
 geom_histogram() +
 xlab("Sample") +
 ylab("Count") +
 ggtitle("GGplot Theme",subtitle = "Testing a ggplot theme for OZminerals") +
 oz_style()
}
